requirements.txt
django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
celery==5.4.0
redis==5.0.7